requirements.txt
streamlit
pandas
ta
selenium
beautifulsoup4
requests
python-dotenv
streamlit-authenticator
chromedriver-autoinstaller
